requirements.txt
playwright>=1.40.0
openai>=1.0.0
pillow>=10.0.0
python-dotenv>=1.0.0
sniffio>=1.3.0
numpy>=1.24.0
captcha-recognizer>=1.0.2
pytweening>=1.2.0